Move session storage to a separate module so that it's easier to understand MimicCore's role.
