+++ mimic/rest/auth_api.py
@@ -37,7 +37,7 @@ class AuthApi(object):
         content = json.loads(request.content.read())
         # tenant_id = content['auth'].get('tenantName', None)
         credentials = content['auth']['passwordCredentials']
-        session = self.core.session_for_username_password(
+        session = self.core.sessions.session_for_username_password(
             credentials['username'], credentials['password'],
             content['auth'].get('tenantName', None),
         )
@@ -68,7 +68,7 @@ class AuthApi(object):
         """
         # FIXME: TEST
         request.setResponseCode(301)
-        session = self.core.session_for_tenant_id(tenant_id)
+        session = self.core.sessions.session_for_tenant_id(tenant_id)
         return json.dumps(dict(user=dict(id=session.username)))
 
     @app.route('/v2.0/RAX-AUTH/impersonation-tokens', methods=['POST'])
@@ -82,7 +82,7 @@ class AuthApi(object):
         expires_in = content['RAX-AUTH:impersonation']['expire-in-seconds']
         username = content['RAX-AUTH:impersonation']['user']['username']
 
-        session = self.core.session_for_impersonation(username, expires_in)
+        session = self.core.sessions.session_for_impersonation(username, expires_in)
         return json.dumps({"access": {
             "token": {"id": session.token,
                       "expires": format_timestamp(session.expires)}
@@ -97,7 +97,7 @@ class AuthApi(object):
         # FIXME: TEST
         request.setResponseCode(200)
         prefix_map = {}
-        session = self.core.session_for_token(token_id)
+        session = self.core.sessions.session_for_token(token_id)
         return json.dumps(get_endpoints(
             session.tenant_id,
             entry_generator=lambda tenant_id: list(
